+++ src/test/java/ru/msugrobov/repositories/WalletRepositoryTest.java
@@ -6,7 +6,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import ru.msugrobov.entities.*;
-import ru.msugrobov.exceptions.IdAlreadyExistsException;
 import ru.msugrobov.exceptions.IdNotFoundException;
 import ru.msugrobov.exceptions.PlayerIdAlreadyExistsException;
 
@@ -18,8 +17,7 @@ import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.assertj.core.api.Assertions.*;
 
 public class WalletRepositoryTest {
 
@@ -31,17 +29,20 @@ public class WalletRepositoryTest {
 
     @BeforeEach
     public void initEach() {
-        Player initPlayer = new Player(1, "Maxim", "Sugrobov",
+        Player initPlayer = new Player("Maxim", "Sugrobov",
                 "Max", "Pass", Role.ADMIN);
-        this.initWallet = new Wallet(1, initPlayer.getId(), new BigDecimal(10000));
         this.testPlayerRepository.create(initPlayer);
+        initPlayer.setId(this.testPlayerRepository.readByLogin("Max").getId());
+        this.initWallet = new Wallet(initPlayer.getId(), new BigDecimal(10000));
+
         this.testRepository.create(initWallet);
+        initWallet.setId(this.testRepository.readByPlayerId(initPlayer.getId()).getId());
     }
 
     @AfterEach
     public void cleanUpEach() {
-        String DELETE_ALL_WALLETS = "DELETE FROM wallets";
-        String DELETE_ALL_PLAYERS = "DELETE FROM players";
+        String DELETE_ALL_WALLETS = "DELETE FROM wallets WHERE player_id <> '1'";
+        String DELETE_ALL_PLAYERS = "DELETE FROM players WHERE login <> 'admin'";
         try (Connection connection = DBconnection.getConnection()) {
             PreparedStatement statementClearWallets = connection.prepareStatement(DELETE_ALL_WALLETS);
             statementClearWallets.executeUpdate();
@@ -55,44 +56,33 @@ public class WalletRepositoryTest {
     @Test
     @DisplayName("Test for creating new wallet")
     public void createWalletTest(){
-        Player playerForWalletCreation = new Player(2, "firstName", "lastName",
+        Player playerForWalletCreation = new Player("firstName", "lastName",
                 "Login", "Password", Role.USER);
-        Wallet wallet = new Wallet(2, 2, new BigDecimal(10000));
         this.testPlayerRepository.create(playerForWalletCreation);
+        Wallet wallet = new Wallet(this.testPlayerRepository.readByLogin("Login").getId(), new BigDecimal(10000));
 
         this.testRepository.create(wallet);
 
         walletRepositoryAssertion.assertThat(this.testRepository).isNotNull();
-        walletRepositoryAssertion.assertThat(this.testRepository.readAll()).hasSize(2);
-        walletRepositoryAssertion.assertAll();
-    }
-
-    @Test
-    @DisplayName("Test for creating new wallet with existing id")
-    public void createWalletWithExistingIdTest(){
-        Wallet walletWithExistingId = new Wallet(1, 1, new BigDecimal(10000));
-
-        walletRepositoryAssertion.assertThatThrownBy(() -> this.testRepository.create(walletWithExistingId))
-                .isInstanceOf(IdAlreadyExistsException.class).hasMessageContaining("id");
-        walletRepositoryAssertion.assertThat(this.testRepository.readAll()).hasSize(1);
+        walletRepositoryAssertion.assertThat(this.testRepository.readAll()).hasSize(3);
         walletRepositoryAssertion.assertAll();
     }
 
     @Test
     @DisplayName("Test for creating new wallet with existing playerId")
     public void createWalletWithExistingPlayerIdTest(){
-        Wallet walletWithExistingPlayerId = new Wallet(2, 1, new BigDecimal(10000));
+        Wallet walletWithExistingPlayerId = new Wallet(1, new BigDecimal(10000));
 
         walletRepositoryAssertion.assertThatThrownBy(() -> this.testRepository.create(walletWithExistingPlayerId))
                 .isInstanceOf(PlayerIdAlreadyExistsException.class).hasMessageContaining("playerId");
-        walletRepositoryAssertion.assertThat(this.testRepository.readAll()).hasSize(1);
+        walletRepositoryAssertion.assertThat(this.testRepository.readAll()).hasSize(2);
         walletRepositoryAssertion.assertAll();
     }
 
     @Test
     @DisplayName("Test for reading wallet info by id")
     public void readTest() {
-        Wallet newWallet = this.testRepository.readById(1);
+        Wallet newWallet = this.testRepository.readById(initWallet.getId());
         assertThat(newWallet).usingRecursiveComparison().isEqualTo(this.initWallet);
     }
 
@@ -106,7 +96,7 @@ public class WalletRepositoryTest {
     @Test
     @DisplayName("Test for reading wallet info by player id")
     public void readByPlayerIdTest() {
-        Wallet newWallet = this.testRepository.readByPlayerId(1);
+        Wallet newWallet = this.testRepository.readByPlayerId(initWallet.getPlayerId());
         assertThat(newWallet).usingRecursiveComparison().isEqualTo(this.initWallet);
     }
 
@@ -120,17 +110,19 @@ public class WalletRepositoryTest {
     @Test
     @DisplayName("Test for updating wallet")
     public void updateTest() {
-        Wallet updatedWallet = new Wallet(1, 1, new BigDecimal(20000));
-        this.testRepository.update(1, updatedWallet);
-        assertThat(this.testRepository.readById(1)).usingRecursiveComparison().isEqualTo(updatedWallet);
+        Wallet updatedWallet = new Wallet(initWallet.getId(), initWallet.getPlayerId(), new BigDecimal(20000));
+        this.testRepository.update(initWallet.getId(), updatedWallet);
+        assertThat(this.testRepository.readById(initWallet.getId()))
+                .usingRecursiveComparison().isEqualTo(updatedWallet);
     }
 
     @Test
     @DisplayName("Test for updating wallet with wrong parameters")
     public void updateWithWrongParametersTest() {
         Wallet updatedWallet = new Wallet(2, 2, new BigDecimal(20000));
-        this.testRepository.update(1, updatedWallet);
-        assertThat(this.testRepository.readById(1)).usingRecursiveComparison().isNotEqualTo(updatedWallet);
+        this.testRepository.update(initWallet.getId(), updatedWallet);
+        assertThat(this.testRepository.readById(initWallet.getId()))
+                .usingRecursiveComparison().isNotEqualTo(updatedWallet);
     }
 
     @Test
@@ -144,8 +136,8 @@ public class WalletRepositoryTest {
     @Test
     @DisplayName("Test for deleting wallet")
     public void deleteTest() {
-        this.testRepository.delete(1);
-        assertThat(this.testRepository.readAll()).hasSize(0);
+        this.testRepository.delete(initWallet.getId());
+        assertThat(this.testRepository.readAll()).hasSize(1);
     }
 
     @Test
@@ -158,13 +150,9 @@ public class WalletRepositoryTest {
     @Test
     @DisplayName("Test for reading all wallets from storage")
     public void readAllTest() {
-        Wallet newTestWallet = new Wallet(2, 2, new BigDecimal(10000));
-        Player playerForWalletCreation = new Player(2, "firstName", "lastName",
-                "Login", "Password", Role.USER);
-        this.testPlayerRepository.create(playerForWalletCreation);
-        this.testRepository.create(newTestWallet);
+        Wallet adminWallet = new Wallet(1, 1, new BigDecimal(10000));
+        this.storage.add(adminWallet);
         this.storage.add(initWallet);
-        this.storage.add(newTestWallet);
 
         List<Wallet> allWalletsInStorage = this.testRepository.readAll();
